장바구니 코드 리팩토링 #98
