+++ src/Type/TypeInterface.tsx
@@ -32,3 +32,16 @@ export interface ProducToShoppingCart {
   productId: number; // 상품아이디
 }
 //디테일페이지에서 상품 정보 보낼때 타입
+
+//장바구니에서 필요한 타입
+export interface TypeCart {
+  id: string;
+  category: string; //카테고리
+  img: string; //이미지
+  artist: string; //아티스트
+  title: string; //상품이름
+  price: number; //가격
+  quantity: number; // 선택된수량
+  selected: boolean; // 체크박스 선택 T/F
+  productId: number; //제품고유id
+}
