+++ src/components/Cart/TopSide.tsx
@@ -1,0 +1,40 @@
+import * as S from '../../styledComponent/styledCart/StCart';
+
+const TopSide = () => {
+  return (
+    <S.CartTitle>
+      <span className="span1">장바구니</span>
+      <S.Process>
+        <span>장바구니</span>
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="16"
+          height="16"
+          viewBox="0 0 16 16"
+          fill="none"
+        >
+          <path
+            d="M6.00022 12C5.99991 11.8442 6.05416 11.6932 6.15355 11.5733L9.14022 7.99996L6.26022 4.41996C6.20484 4.35177 6.16348 4.27331 6.13853 4.18908C6.11358 4.10485 6.10551 4.01653 6.11481 3.92917C6.1241 3.84182 6.15056 3.75717 6.19268 3.68008C6.2348 3.60298 6.29173 3.53498 6.36022 3.47996C6.42841 3.42458 6.50687 3.38323 6.5911 3.35828C6.67532 3.33332 6.76365 3.32526 6.851 3.33455C6.93836 3.34384 7.02301 3.37031 7.1001 3.41242C7.17719 3.45454 7.2452 3.51148 7.30022 3.57996L10.5202 7.57996C10.6183 7.69925 10.6719 7.84888 10.6719 8.00329C10.6719 8.15771 10.6183 8.30734 10.5202 8.42663L7.18688 12.4266C7.13092 12.4941 7.06218 12.5499 6.98462 12.5908C6.90705 12.6317 6.82218 12.6569 6.73486 12.665C6.64755 12.673 6.5595 12.6638 6.47576 12.6378C6.39203 12.6117 6.31425 12.5694 6.24688 12.5133C6.17041 12.4513 6.10862 12.3731 6.06597 12.2843C6.02333 12.1955 6.00087 12.0984 6.00022 12Z"
+            fill="#999999"
+          />
+        </svg>
+        <span>주문결제</span>
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="16"
+          height="16"
+          viewBox="0 0 16 16"
+          fill="none"
+        >
+          <path
+            d="M6.00022 12C5.99991 11.8442 6.05416 11.6932 6.15355 11.5733L9.14022 7.99996L6.26022 4.41996C6.20484 4.35177 6.16348 4.27331 6.13853 4.18908C6.11358 4.10485 6.10551 4.01653 6.11481 3.92917C6.1241 3.84182 6.15056 3.75717 6.19268 3.68008C6.2348 3.60298 6.29173 3.53498 6.36022 3.47996C6.42841 3.42458 6.50687 3.38323 6.5911 3.35828C6.67532 3.33332 6.76365 3.32526 6.851 3.33455C6.93836 3.34384 7.02301 3.37031 7.1001 3.41242C7.17719 3.45454 7.2452 3.51148 7.30022 3.57996L10.5202 7.57996C10.6183 7.69925 10.6719 7.84888 10.6719 8.00329C10.6719 8.15771 10.6183 8.30734 10.5202 8.42663L7.18688 12.4266C7.13092 12.4941 7.06218 12.5499 6.98462 12.5908C6.90705 12.6317 6.82218 12.6569 6.73486 12.665C6.64755 12.673 6.5595 12.6638 6.47576 12.6378C6.39203 12.6117 6.31425 12.5694 6.24688 12.5133C6.17041 12.4513 6.10862 12.3731 6.06597 12.2843C6.02333 12.1955 6.00087 12.0984 6.00022 12Z"
+            fill="#999999"
+          />
+        </svg>
+        <span>주문완료</span>
+      </S.Process>
+    </S.CartTitle>
+  );
+};
+
+export default TopSide;
